requirements.txt
opencv-python>=4.7.0.72
numpy>=1.24
selenium>=4.10
webdriver-manager>=4.0
PyYAML>=6.0
pydantic>=2.5
pyautogui>=0.9.53
pygetwindow>=0.0.9

pytest>=7.0
pillow>=10.0

pynput>=1.7

jinja2>=3.1
